Create API Deploy Github Action (#300)
README is still in progress.
When a branch is merged into master and something in the API folder has changed, this action will deploy the API and the indexer.

Google auth and app-engine packages used in this file is locked to a specific version instead of `get latest`. I prefer to use locked versions for things auth related libs but we wont get potential security updates, and god knows we wont actually update versions once this goes live.