Fixed OpenSea indexer + added logging for monitoring
Needed to add the `include_orders=true` query string, which we were
depending on to fill out orders (and prices) for item sales.

This will now properly pull the correct prices for items on sale, and
add some logging that is necessary to ensure that we can monitor the
health of this service in production.