Attempt to clarify how env variables, secrets, and defaults all work (#319)
Now you can specify accessing an ENV variable, a GCP secret, or a default value if none of those are set. To access a GCP secret, you set either an ENV or a default that begins with "secret://path-to-gcp-secret" and it will be set by fetching that secret from GCP.

This standardizes the way we're accessing all configuration variables so hopefully there is no mystery how these things are initialized.